Project1_ChatBot(Internal DB&Online))/requirements.txt
python-dotenv ==1.1.0
streamlit==1.44.0
tiktoken==0.9.0
protobuf==5.29.3
langchain==0.3.20
langchain-community==0.3.0
langchain-text-splitters==0.3.6
langchain-core==0.3.41
langchain-experimental==0.3.0
langchain-google-genai ==2.0.8
InstructorEmbedding==1.0.1
faiss_cpu==1.10.0
torch==2.2.2
sentence_transformers==4.0.1
google-genai==1.8.0